+++ docs/reference/changelog.rst
@@ -4,6 +4,11 @@ Changelog
 See the `Releases page`_ on GitHub for a complete list of commits that are
 included in each version.
 
+2.1.1 (2025-Feb-11)
+-------------------
+- Fix a bug where instances with underscores in their names would fail to
+  launch.
+
 2.2.0 (2025-Jan-16)
 -------------------
 - ``hookutil.py`` now available for dependent projects to clean up lxd